pyproject.toml
[project]
name = "prs"
version = "2.5.63"
description = "Planning Referral System corporate application"
authors = [{ name = "Ashley Felton", email = "ashley.felton@dbca.wa.gov.au" }]
license = "Apache-2.0"
requires-python = ">=3.13,<4.0"
dependencies = [
  "django==5.2.1",
  "psycopg[binary,pool]==3.2.9",
  "dbca-utils==2.1.3",
  "python-dotenv==1.1.0",
  "dj-database-url==2.3.0",
  "gunicorn==23.0.0",
  "django-extensions==3.2.3",
  "django-crispy-forms==2.3",
  "django-reversion==5.1.0",
  "django-taggit==6.1.0",
  "unidecode==1.4.0",
  "pillow==11.2.1",
  "python-magic==0.4.27",
  "lxml[html-clean]==5.4.0",
  "xmltodict==0.14.2",
  "extract-msg==0.54.1",
  "webtemplate-dbca==1.8.2",
  "geojson==3.2.0",
  "typesense==1.0.3",
  "pdfminer-six==20250506",
  "docx2txt==0.9",
  "celery==5.5.2",
  "django-celery-results==2.5.1",
  "pygeopkg==0.1.3",
  "whitenoise[brotli]==6.9.0",
  "django-crum==0.7.9",
  "sentry-sdk[django]==2.29.1",
  "crispy-bootstrap5==2025.4",
  "redis==5.2.1",
  "xlsxwriter==3.2.3",
  "django-storages[azure]==1.14.6",
]

[dependency-groups]
dev = [
  "ipython>=8.31.0",
  "ipdb>=0.13.13",
  "mixer>=7.2.2",
  "pre-commit>=4.0.1",
  "coverage>=7.6.10",
]

[tool.ruff]
line-length = 140
indent-width = 4

[tool.ruff.lint]
ignore = ["E265", "E501", "E722"]

# Reference: https://www.djlint.com/docs/configuration/
[tool.djlint]
profile = "django"
